Playstream QNA - Move all types to their own files
